refactor: move event-conf enrichment to daemon-common for all Spring Boot daemons
Extract EventConfEnrichmentService from Alarmd's EventEnrichmentConfiguration into
daemon-common so all Spring Boot daemons share a single enrichment implementation.

- Add EventConfEnrichmentService (@Service) to daemon-common with full JDBC loading
  from eventconf_sources/eventconf_events, enrichEvent() for alarm-data/severity/logmsg,
  and all parameter-expansion helpers (expandParms, convertAlarmData, etc.)
- Add opennms-config dependency to daemon-common/pom.xml with standard ServiceMix
  and legacy-framework exclusions
- Simplify EventEnrichmentConfiguration in daemon-boot-alarmd to a thin @Bean that
  wraps AlarmPersisterImpl with EventConfEnrichmentService; remove all inline logic
  and the direct opennms-config dependency (now provided transitively via daemon-common)
- Add EventIpcManagerEnrichingWrapper to daemon-boot-eventtranslator: a full
  EventIpcManager decorator that calls enrichEvent() on all sendNow/sendNowSync/send
  paths before forwarding to the real Kafka manager
- Update EventTranslatorBootConfiguration to inject EventConfEnrichmentService and
  give EventTranslator the enriching wrapper so translated events carry alarm-data
  before reaching Alarmd

E2E test result: 13/13 passed (trap → Minion → Kafka → Trapd → EventTranslator →
Alarmd → PostgreSQL alarm creation and clearing verified)
